backend/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.8.0
asgiref==3.8.1
assemblyai==0.36.0
astunparse==1.6.3
attrs==24.3.0
audioread==3.0.1
babel==2.16.0
bangla==0.0.2
blinker==1.9.0
blis==1.2.0
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.1
coqpit==0.0.17
cryptography==45.0.4
cycler==0.12.1
cymem==2.0.11
Cython==3.0.11
datasets==3.2.0
dateparser==1.1.8
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
Django==5.0.2
django-cors-headers==4.3.1
django-filter==23.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
docopt==0.6.2
einops==0.8.0
elevenlabs==1.50.3
encodec==0.1.1
et_xmlfile==2.0.0
filelock==3.17.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==25.1.21
fonttools==4.55.4
frozenlist==1.5.0
fsspec==2024.9.0
g2pkk==0.1.2
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.69.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.1
gruut-lang-en==2.0.1
gruut-lang-es==2.0.1
gruut-lang-fr==2.0.2
h11==0.14.0
h5py==3.12.1
hangul-romanize==0.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
inflect==7.5.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonlines==1.2.0
keras==3.8.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.44.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.4.1
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
mutagen==1.47.0
namex==0.0.8
networkx==2.8.8
nltk==3.9.1
num2words==0.5.14
numba==0.61.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.60.0
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.14.0
outcome==1.3.0.post0
packaging==24.2
pandas==1.5.3
pillow==10.4.0
platformdirs==4.3.6
playsound==1.3.0
pooch==1.8.2
preshed==3.0.9
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
psycopg2-binary==2.9.10
pyarrow==19.0.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pydub==0.25.1
pygame==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
pynndescent==0.5.13
pyparsing==3.2.1
pypinyin==0.53.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.16.1
pysbd==0.3.4
PySocks==1.7.1
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
selenium==4.29.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soundfile==0.13.0
soxr==0.5.0.post1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.12.0
SQLAlchemy==2.0.36
sqlparse==0.5.3
srsly==2.5.1
SudachiDict-core==20241021
SudachiPy==0.6.10
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
thinc==8.3.4
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.5.1
torchaudio==2.5.1
tqdm==4.67.1
trainer==0.0.36
transformers==4.48.1
trio==0.29.0
trio-websocket==0.12.2
TTS==0.22.0
typeguard==4.4.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
umap-learn==0.5.7
Unidecode==1.3.8
urllib3==2.3.0
wasabi==1.1.3
Wave==0.0.2
weasel==0.4.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.2.1
xxhash==3.5.0
yarl==1.18.3
